requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
api==0.0.7
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.26.89
botocore==1.29.89
brotlipy==0.7.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.6.2
entrypoints==0.4
Flask==2.2.3
Flask-HTTPAuth==4.7.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
nose==1.3.7
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipreqs==0.4.13
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.4
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
smmap==5.0.0
streamlit==1.21.0
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
transformers==4.26.1
tzdata==2023.3
tzlocal==4.3
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wincertstore==0.2
yarg==0.1.9
yarl==1.9.1
zipp==3.15.0
pytesseract==0.3.10
pdf2image==1.16.3
